requirements.txt
alembic==1.7.6
aniso8601==9.0.1
async-timeout==4.0.2
bidict==0.21.4
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.2.0
email-validator==1.1.3
Faker==13.3.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-redis==0.4.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-Simple-Serializer==1.1.3
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.0
importlib-resources==5.4.0
infinity==1.5
intervals==0.9.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.1
python-socketio==5.5.1
pytz==2021.3
redis==4.2.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
urllib3==1.26.9
validators==0.18.2
Werkzeug==2.0.3
wrapt==1.14.0
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
zipp==3.7.0

celery~=5.2.6